Endpoint de recuperation des evaluations reçues par un utilisateur
