+++ backend/src/models/user.model.js
@@ -5,6 +5,7 @@ const { FieldValue} = require('firebase-admin/firestore');
 //const {getAuth} = require("firebase-admin/auth");
 const { getAuth , createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, sendEmailVerification} = require('firebase/auth');
 const auth = getAuth();
+const TrustSystemService = require('../services/trustSystem.service');
 
 
 const USER_COLLECTION = "users"; // Collection Firestore dédiée aux utilisateurs
@@ -46,94 +47,6 @@ const userSchema = Joi.object({
     isBlocked: Joi.boolean().optional(),
 }).strict();   // Empêche l'ajout de champs non définis
 
-// Fonction de mise à jour du score et de la réputation de l'utilisateur
-const updateUserScore = async (userId) => {
-    try {
-        const userRef = db.collection(USER_COLLECTION).doc(userId);
-        const userDoc = await userRef.get();
-        if (!userDoc.exists) throw new Error("Utilisateur introuvable");
-
-        const userData = userDoc.data();
-        const evaluationsAverage = userData.evaluations.average;
-        console.log(`La moyenne des notations de l'utilisateur est ${evaluationsAverage}`);
-
-        const querySnapshot = await db.collection(TONTINE_COLLECTION).where('membersId', 'array-contains', userId).get();
-        const tontines = [];
-        if (querySnapshot) {
-            querySnapshot.forEach(doc => {
-                tontines.push({...doc.data()});
-            });
-        }
-
-        let totalContributions = 0;
-        let punctualContributions = 0;
-        let contributionsAverage;
-        let successfulTontines = 0;
-        let successfulTontinesAverage;
-
-        if (tontines) {
-            tontines.map((tontine) => {
-                console.log("Tontine: ", tontine);
-                if (tontine.status === "terminée") successfulTontines += 1;
-
-                tontine.tours.map((tour) => {
-                    tour.periodeCotisation.map((periodeCotisation) => {
-                        console.log("Periode de cotisation: ", periodeCotisation);
-                        const date_fin = periodeCotisation.date_fin;
-                        console.log("Date de fin: ", date_fin);
-                        if (periodeCotisation.contributions) {
-                            const userContribution = periodeCotisation.contributions.find((contribtution) => contribtution.memberId === userId);
-
-                            if (userContribution) {
-                                console.log("Contribution de l'utilisateur: ", userContribution)
-                                totalContributions += 1;
-                                if (userContribution.date <= date_fin) punctualContributions += 1;
-                            }
-                        }
-                    })
-                })
-            });
-        }
-
-        if (totalContributions > 4) {
-            contributionsAverage = (punctualContributions / totalContributions) * 5;
-        }else contributionsAverage = 3;
-        console.log("Total des contibutions de l'utilisateur:", totalContributions);
-        console.log("Total des contibutions ponctuelles de l'utilisateur: ", punctualContributions);
-        console.log(`La moyenne des contributions de l'utilisateur est ${contributionsAverage}`);
-
-        if (successfulTontines <= 3) {
-            successfulTontinesAverage = 3
-        }else if (successfulTontines > 3 && successfulTontines <= 5)  {
-            successfulTontinesAverage = 4;
-        }else if (successfulTontines > 5) successfulTontinesAverage = 5;
-        console.log(`La moyenne des tontines reussies de l'utilisateur est ${successfulTontinesAverage}`);
-
-        const newScore = parseFloat(((0.5 * contributionsAverage) + (0.2 * successfulTontinesAverage) + (0.3 * evaluationsAverage)).toFixed(2));
-        console.log(`Le nouveau score de l'utilisateur est ${newScore}`);
-
-        let newReputation;
-        if (newScore < 3) {
-            newReputation = "Membre Risqué";
-        }else if (newScore >= 3 && newScore < 4) {
-            newReputation = "Membre Neutre";
-        }else if (newScore >= 4 && newScore < 5) {
-            newReputation = "Membre Fiable";
-        }else if (newScore === 5) newReputation = "Membre Premium";
-        console.log(`Le nouvelle réputation de l'utilisateur est ${newReputation}`);
-
-        // Mise à jour du score et de la réputation dans Firestore
-        await userRef.update({
-            reputation: newReputation,
-            score: newScore
-        });
-        console.log("Score et réputation mis à jour avec succès", newScore, newReputation);
-
-    }catch (error) {
-        console.error("Impossible de mettre à jour le score et la réputation de l'utilisateur: ", error.message);
-        throw new Error(error.message);
-    }
-}
 
 
 class UserModel {
@@ -468,52 +381,6 @@ class UserModel {
         }
     }
 
-
-    /**
-     * Noter un membre dans une tontine
-     * @param {string} memberId - l'id du membre qui donne la note
-     * @param {number} note - la note donnée
-     * @param {string} comment - le commentaire
-     * @param {string} userId - l'id de l'utilisateur qui reçoit la note
-     * @returns {Promise<Object>} - retourne l'id de l'utilisateur et la note reçue
-     */
-    static async noteMember(memberId, note, comment, userId) {
-        try {
-            const memberRef = db.collection(USER_COLLECTION).doc(memberId);
-            const memberDoc = await memberRef.get();
-            if (!memberDoc.exists) throw new Error("Le membre qui essaie d'envoyer la note est introuvable");
-
-            const userRef = db.collection(USER_COLLECTION).doc(userId);
-            const userDoc = await userRef.get();
-            if (!userDoc.exists) throw new Error("L'utilisateur qui doit recevoir la note est introuvable");
-
-            const userData = userDoc.data();
-            const newEvaluations = [
-                ...userData.evaluations.evaluation || [],
-                {memberId: memberId, note: note, comment: comment || "", date: new Date()}
-            ];
-
-            // On calcule la nouvelle moyenne
-            const newAverage = newEvaluations.reduce((acc, val) => acc + val.note, 0) / newEvaluations.length;
-
-            // Mise à jour du document
-            await userRef.update({
-                evaluations: {
-                    evaluation: newEvaluations,
-                    average: parseFloat(newAverage.toFixed(2))
-                }
-            });
-
-            // Mise à jour du score et de la réputation de l'utilisateur
-            await updateUserScore(userId);
-
-            return {userId: userId, evaluation: {memberId: memberId, note: note, comment: comment}};
-
-        }catch(error) {
-            console.error("Impossible de noter l'utilisateur:", error.message);
-            throw new Error(error.message);
-        }
-    }
 
 }
